requirements.txt
python-telegram-bot[job-queue]~=20.0
SQLAlchemy~=2.0
alembic~=1.13
psycopg2-binary~=2.9
python-dotenv~=1.0
pre-commit
black
httpx~=0.26.0
